pyproject.toml
[tool.poetry]
name = "orders"
version = "0.1.0"
description = ""
authors = ["yoni-j <yonyjacoby@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.9.4"
selenium = "^4.19.0"
selenium-wire = "^5.1.0"
redis = "^5.0.3"
pydantic = "^2.7.0"
flask = "^3.0.3"
google-cloud-pubsub = "^2.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
